fixed #3753: Merging the ends of two node ways produces corrupt way
git-svn-id: http://josm.openstreetmap.de/svn/trunk@2333 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/actions/MergeNodesAction.java
@@ -196,7 +196,9 @@ public class MergeNodesAction extends JosmAction {
                 cmds.add(new ChangeCommand(w, newWay));
             }
         }
-        cmds.add(new DeleteCommand(waysToDelete));
+        if (!waysToDelete.isEmpty()) {
+            cmds.add(new DeleteCommand(waysToDelete));
+        }
         return cmds;
     }
